# swagger.yaml
openapi: 3.0.0
info:
  title: Dark Kitchen API
  description: |
    API de gestion d'une dark kitchen.
    Gere les clients, restaurants, produits, commandes, livreurs et livraisons.
  version: 1.0.0
  contact:
    name: Rayane Haddadi
    email: rayane@test.com
servers:
  - url: http://localhost:3000
    description: Serveur local
security:
  - BearerAuth: []   # applique l'auth JWT sur toutes les routes par defaut
paths:
...
  /api/auth/register:
    post:
      summary: Creer un compte
      tags: [Auth]
      security: []   # exemptee de l'auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Compte cree avec succes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthRegisterReponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/login:
    post:
      summary: Se connecter et obtenir un token JWT
      tags: [Auth]
      security: []   # exemptee de l'auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Connexion reussie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginReponse"
        "401":
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erreur"
...
  /api/auth/resend-confirmation:
    post:
      summary: Renvoyer un email de confirmation de creation de compte
      tags: [Auth]
      security: []   # exemptee de l'auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "admin@darkkitchen.com"
      responses:
        "200":
          description: Email renvoye avec succes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageReponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/reset-password:
    post:
      summary: Envoyer un email de reinitialisation de mot de passe
      tags: [Auth]
      security: []   # exemptee de l'auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "admin@darkkitchen.com"
      responses:
        "200":
          description: Email envoye avec succes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageReponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/update-email:
    put:
      summary: Modifier l'email de l'utilisateur connecte
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "nouveau-mail@darkkitchen.com"
      responses:
        "200":
          description: Modification reussie (un email de confirmation est envoye)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageReponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Non connecte
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erreur"
  /api/auth/delete-account:
    delete:
      summary: Supprimer le compte de l'utilisateur connecte
      tags: [Auth]
      responses:
        "200":
          description: Compte supprime avec succes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageReponse"
        "401":
          description: Non connecte
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erreur"
        "500":
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erreur"
  /api/auth/users/{id}/password:
    put:
      summary: Changer le mot de passe d'un utilisateur par son ID (admin)
      tags: [Auth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: UUID de l'utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: "NouveauMotDePasse123"
      responses:
        "200":
          description: Mot de passe mis a jour avec succes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageReponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Non connecte
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erreur"
...
  /api/clients:
    get:
      summary: Lister tous les clients
      tags: [Clients]
      responses:
        "200":
          description: Liste des clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      summary: Creer un client
      tags: [Clients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Client cree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/clients/{id}:
    get:
      summary: Recuperer un client par ID
      tags: [Clients]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Client trouve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Modifier un client
      tags: [Clients]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Client modifie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      summary: Supprimer un client
      tags: [Clients]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Client supprime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageReponse"

  # ═══ RESTAURANTS ════════════════════════════════════════════════════════════
  /api/restaurants:
    get:
      summary: Lister tous les restaurants
      tags: [Restaurants]
      responses:
        "200":
          description: Liste des restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
    post:
      summary: Creer un restaurant
      tags: [Restaurants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantInput"
      responses:
        "201":
          description: Restaurant cree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /api/restaurants/{id}:
    get:
      summary: Recuperer un restaurant par ID
      tags: [Restaurants]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Restaurant trouve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/restaurants/{id}/produits:
    get:
      summary: Lister les produits d'un restaurant
      tags: [Restaurants]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Produits du restaurant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produit"

  # ═══ PRODUITS ═══════════════════════════════════════════════════════════════
  /api/produits:
    get:
      summary: Lister tous les produits
      tags: [Produits]
      responses:
        "200":
          description: Liste des produits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produit"
    post:
      summary: Creer un produit
      tags: [Produits]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProduitInput"
      responses:
        "201":
          description: Produit cree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produit"
  /api/produits/{id}:
    get:
      summary: Recuperer un produit par ID
      tags: [Produits]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Produit trouve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produit"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Modifier un produit
      tags: [Produits]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProduitInput"
      responses:
        "200":
          description: Produit modifie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produit"
    delete:
      summary: Supprimer un produit
      tags: [Produits]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Produit supprime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageReponse"

  # ═══ COMMANDES ══════════════════════════════════════════════════════════════
...
  /api/commandes/{id}:
    get:
      summary: Recuperer une commande par ID
      tags: [Commandes]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Commande trouvee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commande"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenu via POST /api/auth/login.
        A coller dans le champ Authorize en haut de la page sous la forme : Bearer <token>

  # ─── Parametres reutilisables ────────────────────────────────────────────────
...
  responses:
...
    BadRequest:
      description: Donnees invalides ou manquantes
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erreur"

  # ─── Schemas ─────────────────────────────────────────────────────────────────
  schemas:
    AuthInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "admin@darkkitchen.com"
        password:
          type: string
          format: password
          example: "motdepasse123"
...
    AuthRegisterReponse:
      type: object
      properties:
        message:
          type: string
          example: "Compte cree avec succes"
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
    Erreur:
      type: object
      properties:
        message:
          type: string
          example: "Une erreur est survenue"
...